Add files via upload
House price prediction is an important topic of real estate. The literature attempts to derive useful knowledge from historical data of property markets. 
Machine learning techniques are applied to analyze historical property transactions in India to discover useful models for house buyers and sellers. 
Moreover, experiments demonstrate that the Multiple Linear Regression and Random Forest Regressor that is based on accuracy and mean squared error measurement is a competitive approach.
In this project, we have built a machine learning model to predict the house prices .
This project will very helpful for the real estate market.
Buyers are generally not aware of factors that influence the house prices.
 Many problems are faced during buying a house.
The goal of this project is to predict house prices in based on some features such as location, size/area, number of bedrooms, and parking, power backup and etc..
We are using Machine Learning Algorithm to create a predictive model.